🚚 rename: firebase setting 파일 이동
