+++ src/lib/follow/api.ts
@@ -1,4 +1,4 @@
-import { db } from "@/api/database";
+import { db } from "@/shared/firebase";
 import {
   collection,
   deleteDoc,
